:recycle: rename Class
