+++ codeGenerator/src/main/kotlin/org/jmoster/CodeGeneratorApplication.kt
